load files from the local filesystem or fall back to the context
